Add support for passing flags as kwargs to main / start methods.
